Move remaining actions to "notebook actions"
Possibly not the best organizing principle, but at least a little
less ambiguous than what came before. The remaining actions generally
do apply globally, so it's not the worst name.
