+++ src/editor/actions/local-autosave-actions.js
@@ -1,6 +1,6 @@
 import { getLocalAutosave, clearLocalAutosave } from "../tools/local-autosave";
 import { getRevisionID } from "../tools/server-tools";
-import { updateNotebookInfo, updateTitle } from "./actions";
+import { updateNotebookInfo, updateTitle } from "./notebook-actions";
 import { updateIomdContent } from "./editor-actions";
 
 // Check to see if we have an autosaved revision available, if so
